Changing key type of vertex_colors from uint32_t to int64_t
This may be necessary for the C# bindings to generate properly. If this doesn't make it work, I've got another idea to try.

+++ scene/2d/polygon_2d.h
@@ -35,8 +35,6 @@
 #include "core/variant/typed_dictionary.h"
 #include "scene/2d/node_2d.h"
 
-#include <cstdint>
-
 class NavigationPolygon;
 class NavigationMeshSourceGeometryData2D;
 
@@ -45,7 +43,7 @@ class Polygon2D : public Node2D {
 
 	Vector<Vector2> polygon;
 	Vector<Vector2> uv;
-	TypedDictionary<uint32_t, Color> vertex_colors;
+	TypedDictionary<int64_t, Color> vertex_colors;
 	Array polygons;
 	int internal_vertices = 0;
 
@@ -127,8 +125,8 @@ public:
 	void set_color(const Color &p_color);
 	Color get_color() const;
 
-	void set_vertex_colors(const TypedDictionary<uint32_t, Color> &p_colors);
-	TypedDictionary<uint32_t, Color> get_vertex_colors() const;
+	void set_vertex_colors(const TypedDictionary<int64_t, Color> &p_colors);
+	TypedDictionary<int64_t, Color> get_vertex_colors() const;
 
 	void set_texture(const Ref<Texture2D> &p_texture);
 	Ref<Texture2D> get_texture() const;
